requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-import-export==2.8.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
korean-lunar-calendar==0.2.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.2.0
pytz==2022.2.1
PyYAML==6.0
sqlparse==0.4.2
tablib==3.2.1
xlrd==2.0.1
xlwt==1.3.0
django-crispy-forms==1.9.2
fontawesomefree==6.2.1
django-markdownx==3.0.1
